MINE runs
MINE+GAN runs, but there are clearly many issues.

Visualization: it's clear it would be cleaner and clearer to move visualization out of the routine loops. Possibly into a model method.
Model arguments: it's difficult to build noise inside the build routines, as you must have a build plugin to make noise. This will be moved to a model method.
Results: different routines have different results, and these are easily overwritten.
Losses cannot be added to, it seems like this would be useful.
Model arguments are still confusing, and there's no top-level control available, all control is derived from individual routines.
Routines return tensors, but these can't often be used by other routines as the variable is stale if an update happened afterwards..